feat ✨: Update dependencies and add ROCm support
This commit updates the dependencies to use a specific version of NumPy (1.26.4) and adds ROCm support for PyTorch and Onnxruntime. The changes include updating the `pyproject.toml` file to reflect these updates, adding new scripts in the `[tool.pdm.scripts]` section for starting web UI and API services, and modifying the `[tool.pdm.source]` sections to include ROCm-specific packages. These changes ensure compatibility with ROCm hardware and improve the overall performance and reliability of the ASR pipeline.